Refactor persistence logic out of db.rs to storage/index_persistence
Extracted `PersistenceTracker`, background thread management, and index persistence/loading functions from `src/db.rs` into `src/storage/index_persistence/` submodules (`tracker`, `operations`, `background`).

This reduces the complexity of `GallifreyDB` (the God Object) and improves cohesion by keeping persistence details within the storage module.

- Moved `PersistenceTracker` to `src/storage/index_persistence/tracker.rs`.
- Moved `persist_*` and `load_*` functions to `src/storage/index_persistence/operations.rs`.
- Moved `spawn_background_persistence_thread` to `src/storage/index_persistence/background.rs`.
- Updated `GallifreyDB` to delegate to these functions.
- `load_vector_indexes` now takes `&CurrentStorage` instead of `&GallifreyDB` to break circular dependency.
